Java File Camparator or just to compare files
this java netbeans project is a desktop application that compares files
into two different folders and gets the list of files found to be the
same and user can choose to delete the files.

This is designed to remove the files that one already has so as to save
space like for instance i used to download videos  but i would change
location the i realised that i could have downloaded the same video but
i dint know how to compare or check .

I hope you will use this to help you learn more on using the fileI/O
class in java

usage :import using neatbeans and be sure that you have java 8. as i
have employed lambda expressions.Please do note that this is just for
educational reasons feel free to use th code for what ever.